+++ git-index/tests/index/file/mod.rs
@@ -1,4 +1,3 @@
 mod access;
-mod initialize;
 mod read;
 mod write;
