+++ git-index/tests/index/init.rs
@@ -1,10 +1,10 @@
-// use git_index::verify::extensions::no_find;
-use git_index::{decode, State};
+use git_index::State;
+use git_repository as git;
 use git_repository::prelude::FindExt;
 use git_testtools::scripted_fixture_repo_read_only;
 
 #[test]
-fn tree_to_state() {
+fn tree_to_state() -> crate::Result {
     let fixtures = [
         "make_index/v2.sh",
         "make_index/v2_more_files.sh",
@@ -13,14 +13,13 @@ fn tree_to_state() {
     ];
 
     for fixture in fixtures {
-        let repo_dir = scripted_fixture_repo_read_only(fixture).unwrap();
-        let repo = git_repository::open(&repo_dir).unwrap();
+        let repo_dir = scripted_fixture_repo_read_only(fixture)?;
+        let repo = git::open(&repo_dir)?;
 
-        let tree_id = repo.head_commit().unwrap().tree_id().unwrap();
+        let tree_id = repo.head_commit()?.tree_id()?;
 
-        let expected_state =
-            git_index::File::at(repo_dir.join(".git").join("index"), decode::Options::default()).unwrap();
-        let actual_state = State::from_tree(&tree_id, |oid, buf| repo.objects.find_tree_iter(oid, buf).ok()).unwrap();
+        let expected_state = repo.index()?;
+        let actual_state = State::from_tree(&tree_id, |oid, buf| repo.objects.find_tree_iter(oid, buf).ok())?;
 
         println!("{}\n", fixture);
         actual_state
@@ -36,6 +35,7 @@ fn tree_to_state() {
 
         compare_states(&actual_state, &expected_state, fixture)
     }
+    Ok(())
 }
 
 fn compare_states(actual: &State, expected: &State, fixture: &str) {
@@ -59,6 +59,10 @@ fn compare_states(actual: &State, expected: &State, fixture: &str) {
         fixture
     );
     // assert_eq!(actual.entries(), expected.entries(), "entries mismatch in {}", fixture);
+    assert_eq!(
+        actual.entries().iter().map(|e| e.mode).collect::<Vec<_>>(),
+        expected.entries().iter().map(|e| e.mode).collect::<Vec<_>>()
+    );
     assert_eq!(
         actual.path_backing(),
         expected.path_backing(),
