+++ git-index/tests/index/mod.rs
@@ -1,6 +1,7 @@
 use std::path::{Path, PathBuf};
 
 mod file;
+mod init;
 
 pub fn fixture_index_path(name: &str) -> PathBuf {
     let dir = git_testtools::scripted_fixture_repo_read_only(Path::new("make_index").join(name).with_extension("sh"))
